Move box replication helpers to snapshot product file
